+++ snowpack/snowpackCore/VapourTransport.h
@@ -50,15 +50,15 @@
  * @brief This module contains water vapour transport routines for the 1d snowpack model
  */
 class VapourTransport : public WaterTransport {
-	public: 
+	public:
 		VapourTransport(const SnowpackConfig& cfg);
 		void compTransportMass(const CurrentMeteo& Mdata, double& ql, SnowStation& Xdata, SurfaceFluxes& Sdata, const double& surfaceVaporPressure);
 
-	private: 
-		static void EL_INCID(const size_t &e, int Ie[]);
+	private:
+		static void EL_INCID(const int &e, int Ie[]);
 		static void EL_TEMP( const int Ie[], double Te0[], double Tei[], const std::vector<NodeData> &T0, const double Ti[] );
 		static void EL_RGT_ASSEM(double F[], const int Ie[], const double Fe[]);
-		
+
 		bool compDensityProfile(const CurrentMeteo& Mdata, SnowStation& Xdata,
                               const bool& ThrowAtNoConvergence, double& ql, const double& surfaceVaporPressure);
 		bool sn_ElementKtMatrix(ElementData &Edata, double dt, double T0[ N_OF_INCIDENCES ],
@@ -66,7 +66,7 @@ class VapourTransport : public WaterTransport {
 								const std::vector<double> factor_, const int index);
 		void neumannBoundaryConditions(double Se[ N_OF_INCIDENCES ][ N_OF_INCIDENCES ],
                                        double Fe[ N_OF_INCIDENCES ],
-                                       double& X);								                             	 		
+                                       double& X);
 		void compSurfaceSublimation(const CurrentMeteo& Mdata, double& ql, SnowStation& Xdata, SurfaceFluxes& Sdata);
 		void LayerToLayer(const CurrentMeteo& Mdata, SnowStation& Xdata, SurfaceFluxes& Sdata, double& ql, const double& surfaceVaporPressure);
